Remise de area.py et geometry.py dans RULEngine
