+++ ai/STA/Action/Action.py
@@ -12,7 +12,7 @@ class Action:
     """
     def __init__(self, p_game_state):
         """
-            :param p_info_manager: référence vers l'InfoManager
+            :param p_game_state: L'état courant du jeu.
         """
         assert(isinstance(p_game_state, GameState))
         self.game_state = p_game_state
