+++ ai/STA/Action/GrabBall.py
@@ -2,7 +2,7 @@
 from .Action import Action
 from ...Util.types import AICommand
 from RULEngine.Util.Pose import Pose
-from ai.Util.geometry import get_angle
+from RULEngine.Util.geometry import get_angle
 from RULEngine.Util.constant import PLAYER_PER_TEAM
 
 __author__ = 'Robocup ULaval'
@@ -18,7 +18,7 @@ class GrabBall(Action):
     """
     def __init__(self, p_game_state, p_player_id):
         """
-            :param p_info_manager: référence vers l'InfoManager
+            :param p_game_state: L'état courant du jeu.
             :param p_player_id: Identifiant du joueur qui prend le contrôle de la balle
         """
         Action.__init__(self, p_game_state)
