+++ ai/STA/Action/Idle.py
@@ -14,7 +14,7 @@ class Idle(Action):
     """
     def __init__(self, p_game_state, p_player_id):
         """
-            :param p_info_manager: référence vers l'InfoManager
+            :param p_game_state: L'état courant du jeu.
             :param p_player_id: Identifiant du joueur qui s'arrête
         """
         Action.__init__(self, p_game_state)
