+++ ai/STA/Tactic/CoverZone.py
@@ -5,11 +5,11 @@ from ai.STA.Tactic.tactic_constants import Flags
 from ai.STA.Action.GoBetween import GoBetween
 from ai.STA.Action.MoveTo import MoveTo
 from ai.STA.Action.Idle import Idle
-from ai.Util.area import isInsideSquare, stayInsideSquare, stayOutsideGoalArea
+from RULEngine.Util.area import isInsideSquare, stayInsideSquare, stayOutsideGoalArea
 from RULEngine.Util.Pose import Pose
 from RULEngine.Util.Position import Position
 from RULEngine.Util.constant import PLAYER_PER_TEAM, ROBOT_RADIUS
-from ai.Util.geometry import get_angle
+from RULEngine.Util.geometry import get_angle
 
 __author__ = 'RoboCupULaval'
 
@@ -19,7 +19,7 @@ class CoverZone(Tactic):
     méthodes:
         exec(self) : Exécute une Action selon l'état courant
     attributs:
-        info_manager: référence à la façade InfoManager
+        game_state: L'état courant du jeu
         player_id : Identifiant du joueur à qui la tactique est assignée
         y_top : La limite supérieur de la zone
         y_bottom : La limite inférieur de la zone
