+++ ai/STA/Tactic/GoGetBall.py
@@ -7,8 +7,8 @@ from ai.STA.Action.GrabBall import GrabBall
 from ai.STA.Action.Idle import Idle
 from ai.STA.Tactic.tactic_constants import Flags
 
-from ai.Util.area import player_can_grab_ball, player_grabbed_ball
-from ai.Util.geometry import get_distance
+from ai.Util.ball_possession import player_can_grab_ball, player_grabbed_ball
+from RULEngine.Util.geometry import get_distance
 from RULEngine.Util.constant import DISTANCE_BEHIND, PLAYER_PER_TEAM, POSITION_DEADZONE, BALL_RADIUS
 from RULEngine.Util.Pose import Pose
 from RULEngine.Util.Position import Position
@@ -23,7 +23,7 @@ class GoGetBall(Tactic):
     méthodes:
         exec(self) : Exécute une Action selon l'état courant
     attributs:
-        game_state: référence à la façade InfoManager
+        game_state: L'état courant du jeu.
         player_id : Identifiant du joueur auquel est assigné la tactique
         current_state : L'état courant de la tactique
         next_state : L'état suivant de la tactique
